Implementación alertas
Implementación de alertas con SweetAlert2

+++ Programacion II/estetica/src/app/app.ts
@@ -1,3 +1,4 @@
+import Swal from 'sweetalert2';
 import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common'; 
@@ -19,22 +20,56 @@ export class AppComponent {
     this.verTodo(); 
   }
 
-  guardar(){
-    if (!this.servicioActual.nombrePerro || !this.servicioActual.tipoServicio || this.servicioActual.precio <= 0) {
-      alert("Por favor, llena todos los campos correctamente.");
-      return;
+ guardar() {
+  if (!this.servicioActual.nombrePerro || !this.servicioActual.tipoServicio || this.servicioActual.precio <= 0) {
+    Swal.fire({
+      icon: 'error',
+      title: '¡Faltan datos!',
+      text: 'Por favor, llena el nombre, servicio y un precio mayor a 0.',
+      confirmButtonColor: '#3498db'
+    });
+    return;
+  }
+
+  this.servicioActual.id = Date.now();
+  this.esteticaService.agregarServicio({ ...this.servicioActual });
+  this.verTodo();
+
+ 
+  Swal.fire({
+    icon: 'success',
+    title: '¡Perrito registrado!',
+    showConfirmButton: false,
+    timer: 1500
+  });
+
+  this.servicioActual = { id: 0, nombrePerro: '', tipoServicio: '', precio: 0 };
+}
+
+eliminar(id: number) {
+  Swal.fire({
+    title: '¿Eliminar registro?',
+    text: "Esta acción no se puede deshacer",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#d33',
+    cancelButtonColor: '#3085d6',
+    confirmButtonText: 'Sí, borrar',
+    cancelButtonText: 'Cancelar'
+  }).then((result) => {
+
+    if (result.isConfirmed) {
+      this.esteticaService.eliminarServicio(id);
+      this.verTodo();
+      
+      Swal.fire(
+        '¡Borrado!',
+        'El servicio ha sido eliminado correctamente.',
+        'success'
+      );
     }
-    this.servicioActual.id = Date.now();
-    this.esteticaService.agregarServicio({...this.servicioActual});
-    this.verTodo(); 
-    this.servicioActual = { id: 0, nombrePerro: '', tipoServicio: '', precio: 0 };
-  }
-
-  eliminar(id: number) {
-    this.esteticaService.eliminarServicio(id);
-    this.verTodo(); 
-  }
-
+  });
+}
   obtenerTotal(): number {
     return this.listaServicios.reduce((acc, obj) => acc + Number(obj.precio), 0);
   }
